requirements.txt
google-cloud-documentai==2.30.0
google-cloud-storage==2.14.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pypdf==5.1.0
dashscope==1.20.0
